Form template inherited by home

+++ ratings/views.py
@@ -10,13 +10,33 @@ def home(request):
     :param request: Django request object
     :return: rendered homepage
     """
-    if request.POST:
-        rating_id = request.POST['delete']
-        rating = Rating.objects.get(pk=rating_id)
-        rating.delete()
-        return redirect(home)
+    
+    form_object = RatingCreate
 
-    context = {'ratings': Rating.objects.all()}
+    """
+        If post request comes from a delete button:
+            delete.
+        Else:
+            use post reqeust to add data
+    """
+    if request.POST:
+        if 'delete' in request.POST:
+            rating_id = request.POST['delete']
+            rating = Rating.objects.get(pk=rating_id)
+            rating.delete()
+            return redirect(home)
+        else:
+            form = form_object.form_class(request.POST)
+            if form.is_valid():
+                _ = form.save()
+                return redirect(home)
+    
+    form = form_object.form_class()
+
+    context = {
+        'ratings': Rating.objects.all(),
+        'form': form,
+    }
     return render(request, 'home.html', context)
 
 
@@ -25,16 +45,19 @@ class RatingCreate(View):
     form_class = RatingForm
     template_name = 'ratings/rating_form.html'
 
-    def get(self, request):
-        form = self.form_class()
-        return render(request, self.template_name, {'form': form})
+    # def fuckaroo():
+    #     print("fuckaroo")
 
-    def post(self, request):
-        form = self.form_class(request.POST)
-        if form.is_valid():
-            _ = form.save()
-            return redirect(home)
-        return render(request, self.template_name, {'form: form'})
+    # def get(self, request):
+    #     form = self.form_class()
+    #     return render(request, self.template_name, {'form': form})
+
+    # def post(self, request):
+    #     form = self.form_class(request.POST)
+    #     if form.is_valid():
+    #         _ = form.save()
+    #         return redirect(home)
+    #     return render(request, self.template_name, {'form: form'})
 
 
 class RatingEdit(View):
@@ -44,7 +67,7 @@ class RatingEdit(View):
 
     def get(self, request, rating_id):
         rating = Rating.objects.get(pk=rating_id)
-        form = self.form_class(initial={'beer_name': rating.beer_name, 'score': rating.score, 'notes': rating.notes})
+        form = self.form_class(initial={'beer_name': rating.beer_name, 'score': rating.score, 'notes': rating.notes, 'brewery': rating.brewery})
         return render(request, self.template_name, {'form': form})
 
     def post(self, request, rating_id):
